feat: adjust FinanceDate

+++ src/main/java/br/com/devcanoa/finance/api/domain/model/Installments.java
@@ -1,4 +1,4 @@
 package br.com.devcanoa.finance.api.domain.model;
 
-public record Installments(int paid, int total) {
+public record Installments(long paid, long total) {
 }
